+++ test/common/login.js
@@ -1,40 +1,32 @@
+// @flow
 'use strict'
 
 const test = require('ava')
 const proxyquire = require('proxyquire')
 
-const auth0ClientFactoryMock = require('../helpers/auth0-client-factory.js')
 const loginProviderMock = require('../helpers/login-provider.js')
 
 const TEST_SUBJECT = '../../lib/common/login.js'
 
-const CLIENT_ID = 'valid client id'
 const JWT = 'valid jwt'
 
 test.beforeEach((t) => {
-  t.context.auth0ClientFactory = auth0ClientFactoryMock(() => {
-    return Promise.resolve(CLIENT_ID)
-  })
-
-  t.context.usernameLoginProvider = loginProviderMock((clientId, username, password) => {
+  t.context.usernameLoginProvider = loginProviderMock((username, password) => {
     return Promise.resolve(JWT)
   })
 
-  t.context.browserLoginProvider = loginProviderMock((clientId) => {
+  t.context.browserLoginProvider = loginProviderMock(() => {
     return Promise.resolve(JWT)
   })
-
-  t.context.clientName = 'Client Name'
 })
 
 test.cb('login() should return valid jwt from provider', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
+  const login = proxyquire(TEST_SUBJECT, {
     '../login-providers/username.js': t.context.usernameLoginProvider,
     '../login-providers/browser.js': t.context.browserLoginProvider
   })
 
-  commonLogin.login(t.context.clientName)
+  login()
     .then((jwt) => {
       t.is(jwt, JWT)
       t.end()
@@ -45,29 +37,9 @@ test.cb('login() should return valid jwt from provider', (t) => {
     })
 })
 
-test.cb('login() should call auth0ClientFactory with clientName from login', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': auth0ClientFactoryMock((clientName) => {
-      t.is(clientName, t.context.clientName)
-      t.end()
-      return Promise.resolve(CLIENT_ID)
-    }),
-    '../login-providers/username.js': t.context.usernameLoginProvider,
-    '../login-providers/browser.js': t.context.browserLoginProvider
-  })
-
-  commonLogin.login(t.context.clientName)
-    .catch(() => {
-      t.fail()
-      t.end()
-    })
-})
-
 test.cb('login() should create a usernameLoginProvider if all options are passed', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
-    '../login-providers/username.js': loginProviderMock((clientId, username, password) => {
-      t.is(clientId, CLIENT_ID)
+  const login = proxyquire(TEST_SUBJECT, {
+    '../login-providers/username.js': loginProviderMock((username, password) => {
       t.is(username, 'test')
       t.is(password, 'pass')
       t.end()
@@ -76,7 +48,7 @@ test.cb('login() should create a usernameLoginProvider if all options are passed
     '../login-providers/browser.js': t.context.browserLoginProvider
   })
 
-  commonLogin.login(t.context.clientName, {
+  login({
     username: 'test',
     password: 'pass'
   })
@@ -87,10 +59,8 @@ test.cb('login() should create a usernameLoginProvider if all options are passed
 })
 
 test.cb('login() should create a usernameLoginProvider and call login() with null if username is passed in as true', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
-    '../login-providers/username.js': loginProviderMock((clientId, username, password) => {
-      t.is(clientId, CLIENT_ID)
+  const login = proxyquire(TEST_SUBJECT, {
+    '../login-providers/username.js': loginProviderMock((username, password) => {
       t.is(username, null)
       t.is(password, 'pass')
       t.end()
@@ -99,7 +69,7 @@ test.cb('login() should create a usernameLoginProvider and call login() with nul
     '../login-providers/browser.js': t.context.browserLoginProvider
   })
 
-  commonLogin.login(t.context.clientName, {
+  login({
     username: true,
     password: 'pass'
   })
@@ -110,17 +80,16 @@ test.cb('login() should create a usernameLoginProvider and call login() with nul
 })
 
 test.cb('login() should create an browserLoginProvider if no option are passed', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
+  const login = proxyquire(TEST_SUBJECT, {
     '../login-providers/username.js': t.context.usernameLoginProvider,
-    '../login-providers/browser.js': loginProviderMock((clientId, phoneNumber) => {
-      t.is(clientId, CLIENT_ID)
+    '../login-providers/browser.js': loginProviderMock(() => {
+      t.pass()
       t.end()
       return Promise.resolve(JWT)
     })
   })
 
-  commonLogin.login(t.context.clientName)
+  login()
     .catch(() => {
       t.fail()
       t.end()
@@ -128,8 +97,7 @@ test.cb('login() should create an browserLoginProvider if no option are passed',
 })
 
 test('login() should set the correct default for storeJwt option', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
+  const login = proxyquire(TEST_SUBJECT, {
     '../login-providers/username.js': t.context.usernameLoginProvider,
     '../login-providers/browser.js': loginProviderMock((clientId, storeJwt) => {
       t.falsy(storeJwt)
@@ -137,31 +105,29 @@ test('login() should set the correct default for storeJwt option', (t) => {
     })
   })
 
-  return commonLogin.login(t.context.clientName)
+  return login()
 })
 
 test('login() should leave the storeJwt option as true is set to true', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
+  const login = proxyquire(TEST_SUBJECT, {
     '../login-providers/username.js': t.context.usernameLoginProvider,
-    '../login-providers/browser.js': loginProviderMock((clientId, storeJwt) => {
+    '../login-providers/browser.js': loginProviderMock((storeJwt) => {
       t.is(storeJwt, true)
       return Promise.resolve(JWT)
     })
   })
 
-  return commonLogin.login(t.context.clientName, { storeJwt: true })
+  return login({ storeJwt: true })
 })
 
 test('login() should leave the storeJwt option as false is set to false', (t) => {
-  const commonLogin = proxyquire(TEST_SUBJECT, {
-    '../auth0/client-factory.js': t.context.auth0ClientFactory,
+  const login = proxyquire(TEST_SUBJECT, {
     '../login-providers/username.js': t.context.usernameLoginProvider,
-    '../login-providers/browser.js': loginProviderMock((clientId, storeJwt) => {
+    '../login-providers/browser.js': loginProviderMock((storeJwt) => {
       t.is(storeJwt, false)
       return Promise.resolve(JWT)
     })
   })
 
-  return commonLogin.login(t.context.clientName, { storeJwt: false })
+  return login({ storeJwt: false })
 })