+++ test/helpers/login-provider.js
@@ -1,17 +1,9 @@
 'use strict'
 
-const privateVars = new WeakMap()
-
 function loginProviderMock (loginFn) {
   return class LoginProvider {
-    constructor (clientId) {
-      privateVars.set(this, {
-        clientId
-      })
-    }
-
     login (a, b, c) {
-      return loginFn(privateVars.get(this).clientId, a, b, c)
+      return loginFn(a, b, c)
     }
   }
 }