+++ test/login-providers/username.js
@@ -1,3 +1,4 @@
+// @flow
 'use strict'
 
 const test = require('ava')
@@ -24,7 +25,19 @@ test.beforeEach((t) => {
 test.cb('login() should return valid jwt', (t) => {
   const UsernameLoginProvider = proxyquire(TEST_SUBJECT, {
     'inquirer': t.context.inquirer,
-    './login-provider-base.js': t.context.loginProviderBase
+    'aws-sdk': {
+      CognitoIdentityServiceProvider: class {
+        initiateAuth () {
+          return {
+            promise: () => Promise.resolve({
+              AuthenticationResult: {
+                IdToken: JWT
+              }
+            })
+          }
+        }
+      }
+    }
   })
   const usernameLoginProvider = new UsernameLoginProvider(CLIENT_ID)
 
@@ -51,7 +64,19 @@ test.cb('login() should ask for username and password if username and password i
         return memo
       }, {}))
     }),
-    './login-provider-base.js': t.context.loginProviderBase
+    'aws-sdk': {
+      CognitoIdentityServiceProvider: class {
+        initiateAuth () {
+          return {
+            promise: () => Promise.resolve({
+              AuthenticationResult: {
+                IdToken: JWT
+              }
+            })
+          }
+        }
+      }
+    }
   })
   const usernameLoginProvider = new UsernameLoginProvider(CLIENT_ID)
 
@@ -59,7 +84,7 @@ test.cb('login() should ask for username and password if username and password i
     .then(() => {
       t.end()
     })
-    .catch(() => {
+    .catch((e) => {
       t.fail()
       t.end()
     })
@@ -70,7 +95,19 @@ test.cb('login() should should reject if username is not returned from the promp
     'inquirer': inquirerMock((questions) => {
       return Promise.resolve({})
     }),
-    './login-provider-base.js': t.context.loginProviderBase
+    'aws-sdk': {
+      CognitoIdentityServiceProvider: class {
+        initiateAuth () {
+          return {
+            promise: () => Promise.resolve({
+              AuthenticationResult: {
+                IdToken: JWT
+              }
+            })
+          }
+        }
+      }
+    }
   })
   const usernameLoginProvider = new UsernameLoginProvider(CLIENT_ID)
 
@@ -92,7 +129,19 @@ test.cb('login() should should reject if password is not returned from the promp
         username: USERNAME
       })
     }),
-    './login-provider-base.js': t.context.loginProviderBase
+    'aws-sdk': {
+      CognitoIdentityServiceProvider: class {
+        initiateAuth () {
+          return {
+            promise: () => Promise.resolve({
+              AuthenticationResult: {
+                IdToken: JWT
+              }
+            })
+          }
+        }
+      }
+    }
   })
   const usernameLoginProvider = new UsernameLoginProvider(CLIENT_ID)
 
@@ -103,50 +152,6 @@ test.cb('login() should should reject if password is not returned from the promp
     })
     .catch(error => {
       t.deepEqual(error, new Error('Please specify a password.'))
-      t.end()
-    })
-})
-
-test.cb('login() loginProviderBase should contain username and password from prompt and connection should be username based', (t) => {
-  const UsernameLoginProvider = proxyquire(TEST_SUBJECT, {
-    'inquirer': inquirerMock((questions) => {
-      return Promise.resolve({
-        username: USERNAME,
-        password: PASSWORD
-      })
-    }),
-    './login-provider-base.js': loginProviderBaseMock(null, (username, password, connection) => {
-      t.is(username, USERNAME)
-      t.is(password, PASSWORD)
-      t.end()
-      return Promise.resolve(JWT)
-    })
-  })
-  const usernameLoginProvider = new UsernameLoginProvider(CLIENT_ID)
-
-  usernameLoginProvider.login()
-    .catch(() => {
-      t.fail()
-      t.end()
-    })
-})
-
-test.cb('login() should should reject if loginProviderBase returns an error', (t) => {
-  const UsernameLoginProvider = proxyquire(TEST_SUBJECT, {
-    'inquirer': t.context.inquirer,
-    './login-provider-base.js': loginProviderBaseMock(null, (username, password, connection) => {
-      return Promise.reject(new Error('Test error message'))
-    })
-  })
-  const usernameLoginProvider = new UsernameLoginProvider(CLIENT_ID)
-
-  usernameLoginProvider.login()
-    .then(() => {
-      t.fail()
-      t.end()
-    })
-    .catch(error => {
-      t.deepEqual(error, new Error('Test error message'))
       t.end()
     })
 })